WIP: trying to merge two LW components into one, not working yet
